requirements.txt
python-dotenv
requests
streamlit>=1.28.0
